+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 <!--
 
 // Please add your own contribution below inside the Master section, no need to
-// set a version number, that happens during a deploy.
+// set a version number, that happens during a deploy. Thanks!
 //
 // These docs are aimed at users rather than danger developers, so please limit technical
 // terminology in here.